requirements.txt
pandas
flask
python-dotenv
requests
ibm-watsonx-orchestrate
faker
numpy
gunicorn